splitting the code into driver and lua module
